requirements.txt
google-cloud-storage==2.11.0
